build: change project directories

+++ Libs/Console_args/console_args.cpp
@@ -83,7 +83,7 @@ bool handle_cmd_args(const int argc, char* const* argv, const char* format, Args
 void print_help(const char* format)
 {
     printf("OPTIONS:\n");
-    for (int i = 0; i < strlen(format); i++)
+    for (size_t i = 0; i < strlen(format); i++)
     {
         if (format[i] != ':')
         {
